+++ controller/controller.user.js
@@ -16,6 +16,17 @@ const getUser = async (req, res) => {
   }
 };
 
+const deleteUser = async (req, res) => {
+  try {
+    const user = await User.destroy({ where: { id: req.params.user_id } });
+    res.status(200).json({ success: true });
+  } catch (e) {
+    console.log(e);
+    res.status(400).json({ success: false, error: e.message });
+  }
+};
+
 module.exports = {
   getUser,
+  deleteUser,
 };
