+++ lib/lib.Preference.js
@@ -3,7 +3,6 @@ const {
   PreferenceAge,
   PreferencePrice,
   PreferenceGroup,
-  ProductPreference,
   Category,
 } = require("../models");
